"Manage Date Util functions globally"
